+++ notebooks/rapids_scanpy_funcs.py
@@ -18,12 +18,17 @@ import cupy as cp
 import cudf
 import cugraph
 
+import dask
+from cuml.dask.common.part_utils import _extract_partitions
+from cuml.common.memory_utils import with_cupy_rmm
+
 import numpy as np
 import pandas as pd
 import scipy
 import math
 
 from cuml.linear_model import LinearRegression
+from statsmodels import robust
 
 
 def scale(normalized, max_value=10):
@@ -484,3 +489,199 @@ def leiden(adata):
     clusters = pd.Categorical(clusters)
     
     return clusters
+
+@with_cupy_rmm
+def sq_sum_csr_matrix(client, csr_matrix, axis=0):
+    '''
+    Implements sum operation for dask array when the backend is cupy sparse csr matrix
+    '''
+    client = dask.distributed.default_client()
+
+    def __sq_sum(x):
+        x = x.multiply(x)
+        return x.sum(axis=axis)
+
+    parts = client.sync(_extract_partitions, csr_matrix)
+    futures = [client.submit(__sq_sum,
+                             part,
+                             workers=[w],
+                             pure=False)
+               for w, part in parts]
+    objs = []
+    for i in range(len(futures)):
+        obj = dask.array.from_delayed(futures[i],
+                                      shape=futures[i].result().shape,
+                                      dtype=cp.float32)
+        objs.append(obj)
+    return dask.array.concatenate(objs, axis=axis).compute().sum(axis=axis)
+
+
+@with_cupy_rmm
+def sum_csr_matrix(client, csr_matrix, axis=0):
+    '''
+    Implements sum operation for dask array when the backend is cupy sparse csr matrix
+    '''
+    client = dask.distributed.default_client()
+
+    def __sum(x):
+        return x.sum(axis=axis)
+
+    parts = client.sync(_extract_partitions, csr_matrix)
+    futures = [client.submit(__sum,
+                             part,
+                             workers=[w],
+                             pure=False)
+               for w, part in parts]
+    objs = []
+    for i in range(len(futures)):
+        obj = dask.array.from_delayed(futures[i],
+                                      shape=futures[i].result().shape,
+                                      dtype=cp.float32)
+        objs.append(obj)
+    return dask.array.concatenate(objs, axis=axis).compute().sum(axis=axis)
+
+
+def read_with_filter(client,
+                     sample_file,
+                     min_genes_per_cell=200,
+                     max_genes_per_cell=6000,
+                     min_cells = 0,
+                     num_cells=None,
+                     batch_size=50000,
+                     partial_post_processor=None):
+    """
+    Reads an h5ad file and applies cell and geans count filter. Dask Array is
+    used allow partitioning the input file. This function supports multi-GPUs.
+    """
+
+    # Path in h5 file
+    _data = '/X/data'
+    _index = '/X/indices'
+    _indprt = '/X/indptr'
+    _genes = '/var/_index'
+    _barcodes = '/obs/_index'
+
+    @dask.delayed
+    def _read_partition_to_sparse_matrix(sample_file,
+                                         total_cols, batch_start, batch_end,
+                                         min_genes_per_cell=200,
+                                         max_genes_per_cell=6000,
+                                         post_processor=None):
+        with h5py.File(sample_file, 'r') as h5f:
+            indptrs = h5f[_indprt]
+            start_ptr = indptrs[batch_start]
+            end_ptr = indptrs[batch_end]
+
+            # Read all things data and index
+            sub_data = cp.array(h5f[_data][start_ptr:end_ptr])
+            sub_indices = cp.array(h5f[_index][start_ptr:end_ptr])
+
+            # recompute the row pointer for the partial dataset
+            sub_indptrs  = cp.array(indptrs[batch_start:(batch_end + 1)])
+            sub_indptrs = sub_indptrs - sub_indptrs[0]
+
+        # Reconstruct partial sparse array
+        partial_sparse_array = cp.sparse.csr_matrix(
+            (sub_data, sub_indices, sub_indptrs),
+            shape=(batch_end - batch_start, total_cols))
+
+        # TODO: Add barcode filtering here.
+        degrees = cp.diff(partial_sparse_array.indptr)
+        query = ((min_genes_per_cell <= degrees) & (degrees <= max_genes_per_cell))
+        partial_sparse_array = partial_sparse_array[query]
+
+        if post_processor is not None:
+            partial_sparse_array = post_processor(partial_sparse_array)
+
+        return partial_sparse_array
+
+
+    with h5py.File(sample_file, 'r') as h5f:
+        # Compute the number of cells to read
+        indptr = h5f[_indprt]
+        genes = cudf.Series(h5f[_genes], dtype=cp.dtype('object'))
+
+        total_cols = genes.shape[0]
+        max_cells = indptr.shape[0] - 1
+        if num_cells is not None:
+            max_cells = num_cells
+
+    dls = []
+    for batch_start in range(0, max_cells, batch_size):
+        actual_batch_size = min(batch_size, max_cells - batch_start)
+        dls.append(dask.array.from_delayed(
+                   (_read_partition_to_sparse_matrix)
+                   (sample_file,
+                    total_cols,
+                    batch_start,
+                    batch_start + actual_batch_size,
+                    min_genes_per_cell=min_genes_per_cell,
+                    max_genes_per_cell=max_genes_per_cell,
+                    post_processor=partial_post_processor),
+                   dtype=cp.float32,
+                   shape=(actual_batch_size, total_cols)))
+
+    dask_sparse_arr =  dask.array.concatenate(dls)
+    dask_sparse_arr = dask_sparse_arr.persist()
+
+    # Filter by genes (i.e. cell count per gene)
+    gene_wise_cell_cnt = sum_csr_matrix(client, dask_sparse_arr)
+    query = gene_wise_cell_cnt > min_cells
+
+    # Filter genes for var
+    genes = genes[query]
+    genes = genes.reset_index(drop=True)
+
+    query = cp.where(query == True)[0]
+    dask_sparse_arr = dask_sparse_arr[:, query.get()].persist()
+
+    return dask_sparse_arr, genes, query
+
+
+def highly_variable_genes_filter(client,
+                                 data_mat,
+                                 genes,
+                                 n_top_genes=None):
+
+    if n_top_genes is None:
+        n_top_genes = genes.shape[0] // 10
+
+    mean = sum_csr_matrix(client, data_mat, axis=0) / data_mat.shape[0]
+    mean[mean == 0] = 1e-12
+
+    mean_sq = sq_sum_csr_matrix(client, data_mat, axis=0) / data_mat.shape[0]
+    variance = mean_sq - mean ** 2
+    variance *= data_mat.shape[1] / (data_mat.shape[0] - 1)
+    dispersion = variance / mean
+
+    df = pd.DataFrame()
+    df['genes'] = genes.to_array()
+    df['means'] = mean.tolist()
+    df['dispersions'] = dispersion.tolist()
+    df['mean_bin'] = pd.cut(
+        df['means'],
+        np.r_[-np.inf, np.percentile(df['means'], np.arange(10, 105, 5)), np.inf],
+    )
+
+    disp_grouped = df.groupby('mean_bin')['dispersions']
+    disp_median_bin = disp_grouped.median()
+
+    with warnings.catch_warnings():
+        warnings.simplefilter('ignore')
+        disp_mad_bin = disp_grouped.apply(robust.mad)
+        df['dispersions_norm'] = (
+            df['dispersions'].values - disp_median_bin[df['mean_bin'].values].values
+        ) / disp_mad_bin[df['mean_bin'].values].values
+
+    dispersion_norm = df['dispersions_norm'].values
+
+    dispersion_norm = dispersion_norm[~np.isnan(dispersion_norm)]
+    dispersion_norm[::-1].sort()
+
+    if n_top_genes > df.shape[0]:
+        n_top_genes = df.shape[0]
+
+    disp_cut_off = dispersion_norm[n_top_genes - 1]
+    vaiable_genes = np.nan_to_num(df['dispersions_norm'].values) >= disp_cut_off
+
+    return vaiable_genes
